change: パス変数として渡された服用記録IDをバリデーションするように変更

+++ src/main/kotlin/example/presentation/controller/api/medicationrecord/MedicationRecordDeletionApiController.kt
@@ -2,6 +2,7 @@ package example.presentation.controller.api.medicationrecord
 
 import example.application.service.medicationrecord.*
 import example.domain.model.medicationrecord.*
+import example.domain.shared.exception.*
 import example.presentation.shared.usersession.*
 import org.springframework.http.*
 import org.springframework.stereotype.*
@@ -17,6 +18,9 @@ class MedicationRecordDeletionApiController(private val medicationRecordService:
     @PostMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
     fun deleteMedicationRecord(@PathVariable medicationRecordId: MedicationRecordId) {
+        if (!medicationRecordService.isValidMedicationRecordId(medicationRecordId))
+            throw InvalidEntityIdException(medicationRecordId)
+
         medicationRecordService.deleteMedicationRecord(medicationRecordId,
                                                        userSessionProvider.getUserSessionOrElseThrow())
     }
